Change 20091027-hqm-6 by hqm@badtzmaru.home on 2009-10-27 09:23:10 EDT     in /Users/hqm/openlaszlo/trunk6     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: make custom cursors work in swf10, add groundwork for DHTML

New Features:

Bugs Fixed: LPP-8563

Technical Reviewer: max
QA Reviewer: andre
Doc Reviewer: (pending)

Documentation:

Release Notes:

A pair of attributes has been added to the <resource> tag, 'offsetx' and 'offsety'.

The purpose is to allow you to specify an offset of a bitmap resource used for a custom cursor,
so that the mouse will aligh with the proper 'hotspot' in the cursor image. 

Example: set a custom crosshair cursor from a bitmap resource, with an offset of (-8,-8)

<resource name="crosshair" src="cursor_crosshair.png" offsetx="-8" offsety="-8" />
<view height="40" bgcolor="#cccccc" width="100" cursor="crosshair" />


The cursor offset feature currently only works in the swf10 runtime, but will eventually work
in DHTML and SWF8.


Details:
    
This change makes several custom cursor features work in swf10; the
use of bitmap resources, hotspot offsets, and global control of the
whether the handcursor is enabled over clickable views.

It lays the groundwork for bitmap cursors in DHTML, and cursor hotspot
offsets in SWF8.

+ add 'offsetx' and 'offsety' attributes to <resource> tag, so LzMouseKernel can 
implement hotspots for bitmap cursors. These are passed through the the LzResourceLibrary
object in SWF10, DHTML. SWF8 has not yet got a LzResourceLibrary object, so it has
no place to store the offset info for resources.

+ If an offsetx or offsety are found in the LzResourceLibrary for a cursor, LzMouseKernel uses that
delta to offset the cursor position.


Tests:

test/cursor/cursors.lzx in swf10, swf8



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@15069 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c