Change 20100112-maxcarlson-k by maxcarlson@bank on 2010-01-12 11:34:22 PST     in /Users/maxcarlson/openlaszlo/trunk-clean     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fix braino in html tag lookup

Bugs Fixed: LPP-8308 - in IE7 if you select text in an Iframe and then click into an inputtext the text in the iframe is unselected

Technical Reviewer: ptw
QA Reviewer: hminsky

Details: I'm checking this in ahead of review.  

Be more careful when looking to see if a sprite belongs to an html tag.

Tests: test/lfc/legals/keyboardandmouse.lzx?lzr=dhtml.  Also see testcase (lpp-8308.zip) attached to LPP-8308 in IE 7 and the testcase attached to LPP-8375.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@15452 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c